interactive_search_app/requirements.txt
databricks-sdk
streamlit 